Add `CRMS::NewYear` method `choose_rights_prediction`
This is intended to replace one of the nastiest blocks of code encountered in `bin/newyear.pl`
around lines 198-215 and duplicated at 359-376. Much depends on this being comprehensible
and working as intended, so this is an attempt to get a better handle on it.

Currently this new method is not swapped in. I would like to do an A/B comparison between the
two methods of comparing PDD rights predictions.